Sensor unit working as intended with tape- and distance-sensor

+++ sensorenhet/sensorenhet/tapeSensor.h
@@ -16,7 +16,7 @@
 
 extern uint16_t tape_data_done;
 
-int convertToBit(uint16_t data);
+uint16_t convertToBit(uint16_t data);
 
 void calibrateTapeSensor();
 void setOnTape();
